pyproject.toml
[tool.poetry]
name = "cotton_flower_mot"
version = "0.1.0"
description = "Multi-object tracking for cotton flowers."
authors = ["Daniel Petti <djpetti@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10"
kedro = {extras = ["matplotlib", "pandas"], version = "^0.18.1"}
pycvat = {path = "pycvat", develop = false}
pandas = "1.4.3"
loguru = "^0.5.3"
spektral = "^1.2.0"
seaborn = "^0.11.1"
matplotlib = "^3.3.4"
pydantic = "^1.8.1"
scipy = "^1.6.2"
opencv-python = "^4.7.0"
tenacity = "^8.0.1"
tensorflow-datasets = "^4.4.0"
numpy = "^1.26.1"
opencv-contrib-python = "^4.7.0"
tqdm = "^4.64.0"
tensorflow-addons = "^0.19.0"
python-json-logger = "^2.0.4"
onnx2tf = "^1.13.11"
tf2onnx = "^1.16.0"
tabulate = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-mock = "^3.5.1"
pytest-black = "^0.3.12"
black = "^22.3.0"
isort = "^5.7.0"
pre-commit = "^2.11.0"
pytest-snapshot = "^0.5.0"
Faker = "^17.6.0"
pytest-coverage = "^0.0"
PyYAML = "^5.4.1"
pytest-flake8 = "^1.0.7"
pydot = "^1.4.2"
jupyterlab = "^3.4.2"

[tool.poetry.group.dev.dependencies]
ultralytics = "^8.0.208"

[tool.black]
line-length = 79
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.venv
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 79
skip = "('.eggs', '.git', '.venv', 'build', 'dist')"

[tool.kedro]
package_name = "cotton_flower_mot"
project_name = "Cotton Flower MOT"
project_version = "0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
